refactor(jerni-dev-event-server): extract read file and append to file function
